Prepare 0.6.0 release
- Updated redis to 0.24.
- Updated tgbot to 0.20.
- Updated tokio to 1.35.

+++ src/cache.rs
@@ -6,7 +6,7 @@ use tgbot::types::ChatId;
 use tokio::sync::Mutex;
 
 const PREFIX: &str = "telefeed";
-const LIFETIME: usize = 86400 * 7;
+const LIFETIME: u64 = 86400 * 7;
 
 #[derive(Clone)]
 pub struct Cache {
